requirements.txt
pytest
selenium
requests
allure-pytest
python-dotenv
webdriver-manager
